report housekeeping
